fixed issues

+++ DB Scripts PostgreSQL/Scripts/04 - Functions/CN/get_child_node_types_hierarchy.sql
@@ -17,7 +17,7 @@ BEGIN
 	RETURN QUERY
 	WITH tbl AS
 	(
-		SELECT nt.node_type_id AS "id", nt.parent_id, 0 AS "level", nt.name
+		SELECT nt.node_type_id AS "id", nt.parent_id, 0::INTEGER AS "level", nt.name
 		FROM UNNEST(vr_node_type_ids) AS n
 			INNER JOIN cn_node_types AS nt
 			ON nt.application_id = vr_application_id AND nt.node_type_id = n
